Refactor eks config
