+++ internal/cmd/config.go
@@ -421,6 +421,8 @@ func Configure(v *viper.Viper, _ *pflag.FlagSet) {
 	v.SetDefault("cloud.amazon.defaultRegion", "us-west-1")
 	v.SetDefault("cloud.alibaba.defaultRegion", "eu-central-1")
 
+	v.SetDefault("distribution.eks.templateLocation", "./templates/eks")
+
 	v.SetDefault("cloudinfo.endpoint", "")
 
 	// Temporary hook flags
