+++ internal/global/ingresscert/cert_generator.go
@@ -1,4 +1,4 @@
-// Copyright © 2018 Banzai Cloud
+// Copyright © 2019 Banzai Cloud
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,12 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global
+package ingresscert
 
 import (
 	"path/filepath"
 	"sync"
 
+	"github.com/banzaicloud/pipeline/internal/global"
 	"github.com/banzaicloud/pipeline/pkg/crypto/cert"
 	"github.com/banzaicloud/pipeline/secret"
 )
@@ -31,15 +32,15 @@ var certGeneratorOnce sync.Once
 func newCertGenerator() *cert.Generator {
 	var caLoader cert.CALoader
 
-	switch Config.Cluster.Ingress.Cert.Source {
+	switch global.Config.Cluster.Ingress.Cert.Source {
 	case "file":
 		caLoader = cert.NewFileCALoader(
-			filepath.Join(Config.Cluster.Ingress.Cert.Path, "ca.crt.pem"),
-			filepath.Join(Config.Cluster.Ingress.Cert.Path, "ca.key.pem"),
+			filepath.Join(global.Config.Cluster.Ingress.Cert.Path, "ca.crt.pem"),
+			filepath.Join(global.Config.Cluster.Ingress.Cert.Path, "ca.key.pem"),
 		)
 
 	case "vault":
-		caLoader = cert.NewVaultCALoader(secret.Store.Logical, Config.Cluster.Ingress.Cert.Path)
+		caLoader = cert.NewVaultCALoader(secret.Store.Logical, global.Config.Cluster.Ingress.Cert.Path)
 	}
 
 	generator := cert.NewGenerator(cert.NewCACache(caLoader))
